extract utility functions
